Update setup.py
fix for non existing settings file

+++ setup.py
@@ -33,7 +33,8 @@ if removePack:
     print 'Removed files.'
 else:
     dirIn = str(raw_input( "Basic Setup (Input anything other than *advanced* for basic setup):"))
-    os.remove(SETTINGSFILE)
+    if os.path.exists(SETTINGSFILE):
+        os.remove(SETTINGSFILE)
     if not  'dvanced' in dirIn:
         dirApp = os.getcwd()
         dirIn = dirApp
